Make CursorAnimator interface
