+++ editor/src/main/java/io/github/rosemoe/sora/widget/CursorAnimator.java
@@ -23,92 +23,19 @@
  */
 package io.github.rosemoe.sora.widget;
 
-import android.animation.ValueAnimator;
+public interface CursorAnimator {
 
-/**
- * Helper class for cursor animation in editor
- *
- * @author Rosemoe
- */
-class CursorAnimator implements ValueAnimator.AnimatorUpdateListener {
+    void markStartPos();
+    void markEndPos();
 
-    ValueAnimator animatorX;
-    ValueAnimator animatorY;
-    ValueAnimator animatorBgBottom;
-    ValueAnimator animatorBackground;
-    private final CodeEditor editor;
-    private float startX, startY, startSize, startBottom;
-    private long lastAnimateTime;
+    void start();
+    void cancel();
 
-    public CursorAnimator(CodeEditor editor) {
-        this.editor = editor;
-        animatorX = new ValueAnimator();
-        animatorY = new ValueAnimator();
-        animatorBackground = new ValueAnimator();
-        animatorBgBottom = new ValueAnimator();
-    }
+    boolean isRunning();
 
-    public void markStartPos() {
-        var line = editor.getCursor().getLeftLine();
-        float[] pos = editor.mLayout.getCharLayoutOffset(line, editor.getCursor().getLeftColumn());
-        startX = editor.measureTextRegionOffset() + pos[1];
-        startY = pos[0];
-        startSize = editor.mLayout.getRowCountForLine(line) * editor.getRowHeight();
-        startBottom = editor.mLayout.getCharLayoutOffset(line, editor.getText().getColumnCount(line))[0];
-    }
+    float animatedX();
+    float animatedY();
 
-    public boolean isRunning() {
-        return animatorX.isRunning() || animatorY.isRunning() || animatorBackground.isRunning() || animatorBgBottom.isRunning();
-    }
-
-    public void cancel() {
-        animatorX.cancel();
-        animatorY.cancel();
-        animatorBackground.cancel();
-        animatorBgBottom.cancel();
-    }
-
-    public void markEndPosAndStart() {
-        if (!editor.isCursorAnimationEnabled()) {
-            return;
-        }
-        if (isRunning()) {
-            startX = (float) animatorX.getAnimatedValue();
-            startY = (float) animatorY.getAnimatedValue();
-            startSize = (float) animatorBackground.getAnimatedValue();
-            startBottom = (float) animatorBgBottom.getAnimatedValue();
-            cancel();
-        }
-        var duration = 120;
-        if (System.currentTimeMillis() - lastAnimateTime < 100) {
-            return;
-        }
-        var line = editor.getCursor().getLeftLine();
-        animatorX.removeAllUpdateListeners();
-        float[] pos = editor.mLayout.getCharLayoutOffset(editor.getCursor().getLeftLine(), editor.getCursor().getLeftColumn());
-
-        animatorX = ValueAnimator.ofFloat(startX, (pos[1] + editor.measureTextRegionOffset()));
-        animatorY = ValueAnimator.ofFloat(startY, pos[0]);
-        animatorBackground = ValueAnimator.ofFloat(startSize, editor.mLayout.getRowCountForLine(editor.getCursor().getLeftLine()) * editor.getRowHeight());
-        animatorBgBottom = ValueAnimator.ofFloat(startBottom, editor.mLayout.getCharLayoutOffset(line, editor.getText().getColumnCount(line))[0]);
-
-        animatorX.addUpdateListener(this);
-
-        animatorX.setDuration(duration);
-        animatorY.setDuration(duration);
-        animatorBackground.setDuration(duration);
-        animatorBgBottom.setDuration(duration);
-
-        animatorX.start();
-        animatorY.start();
-        animatorBackground.start();
-        animatorBgBottom.start();
-
-        lastAnimateTime = System.currentTimeMillis();
-    }
-
-    @Override
-    public void onAnimationUpdate(ValueAnimator animation) {
-        editor.postInvalidateOnAnimation();
-    }
+    float animatedLineHeight();
+    float animatedLineBottom();
 }
